Replaced NoSeg handler with Task version
Completely removed old, simple NoSeg handler, since it was *too* simple
and a TSS with Not-Present Selectors wouldn't be properly set up.

By having a whole Task to handle it, the Task Switch into the handler
does the job, while the `IRET` re-loads the now-present Selector - and
checks any others.

After doing this though, I tested it in VirtualBox, which made the
Demonstrator GPF! I used the GPF display and the in-built Debugger to
determine that the Error code that VirtualBox `PUSH`ed on the Stack
was... `0000h`! VMware pushes the right value! And it works on real
hardware (Core i7) too.

And Virtual PC just gives up the ghost - the Window silently disappears.

So I `%define`d `System.Present` as `x86.Desc.Type.Present` in
`Demo/Demo.inc`, and used that symbol everywhere I wanted to test the
Not Present code (note there are still references to the original
symbol: they're important!) If you are using VirtualBox or Virtual PC,
you'll want to change the definition to `0`.