Allowed for MBR to write to me
I found a ***nasty*** USB memory stick whose MBR actually *wrote* to the
freshly-loaded VBR without checking that it was a DOS one!

Since I'm not, it kind-of mucked up my code! This commit adds an
embedded area for the USB stick to play with that the code `JMP`s
around.

Another approach would be for my Boot sector to have a BIOS Parameter
Block. To that end I've added the definitions - but I haven't actually
inserted one in the Boot block.