Added Not Present handler to BadStack
If a Stack Segment is marked Not Present, you don't get `#NP`, you get
`#SF`. So I turned (most of) the Not Present handler into a callable
function, and called it from the BadStack handler.
If that works, I just return - with a new Stack to work with!
If it errors, I drop through to the standard Fault handler